requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg[binary]
pgvector
langchain-core
python-dotenv
sentence-transformers
aio-pika